requirements.txt
python-dotenv
tiktoken
flask-cors
openai==0.28.1
Flask
pypdf
langchain==0.0.331
torch
accelerate
bitsandbytes
transformers
sentence_transformers
faiss_cpu